fix(gui): address Codex review on v2 strut, quota aggregation, session bootstrap
- Models: reserve only the sub-agent mode row while /api/v2 loads; the conditional threads row no longer reserves height that vanishes on the default disabled install.
- Provider overview: reuse ProviderCapacityQuota (shared with the aggregate dashboard) so pooled Codex capacity keeps estimate, recovery, current-account, and incomplete-window warnings instead of collapsing to a plain rate-limit bar.
- Dev session bootstrap: serve /opencodex-session from a dedicated minimal document independent of gui/dist, so a fresh Vite checkout mints an origin-bound loopback session without a packaged GUI build.
