+++ gui/src/components/provider-workspace/ProviderOverviewDashboard.tsx
@@ -10,7 +10,6 @@ import type { WorkspaceSections, WorkspaceItem } from "../../provider-workspace/
 import {
   accountQuotaFromReport,
   capacityAggregationFromReport,
-  type CapacityWindowView,
   type ProviderQuotaReportView,
 } from "../../provider-workspace/report";
 import {
@@ -22,10 +21,11 @@ import {
   relativeTimeLabelsFromT,
   type ProviderUsageTotals,
 } from "../../provider-workspace/usage";
-import { maxQuotaUtilisation, type QuotaWindowKey } from "../QuotaBars";
+import { maxQuotaUtilisation } from "../QuotaBars";
 import { ProviderIcon } from "./ProviderRail";
 import { formatProviderDisplayName } from "../../provider-icons";
 import QuotaBars from "../QuotaBars";
+import { ProviderCapacityQuota } from "./ProviderCapacityQuota";
 
 export default function ProviderOverviewDashboard({
   sections,
@@ -233,86 +233,6 @@ export default function ProviderOverviewDashboard({
         </section>
       </div>
     </div>
-  );
-}
-
-function ProviderCapacityQuota({ report, pending }: { report: ProviderQuotaReportView; pending: boolean }) {
-  const t = useT();
-  const { locale } = useI18n();
-  const aggregation = capacityAggregationFromReport(report);
-  const primaryQuota = accountQuotaFromReport(report);
-  const showsAggregate = aggregation?.presentation === "aggregate";
-  const incompleteWindowKeys = new Set<QuotaWindowKey>();
-  const incompleteCustomWindowLabels = new Set<string>();
-  if (showsAggregate && aggregation) {
-    if (aggregation.fiveHour?.incomplete) incompleteWindowKeys.add("fiveHour");
-    if (aggregation.weekly?.incomplete) incompleteWindowKeys.add("weekly");
-    if (aggregation.monthly?.incomplete) incompleteWindowKeys.add("monthly");
-    for (const window of aggregation.customWindows ?? []) {
-      if (window.incomplete) incompleteCustomWindowLabels.add(window.label);
-    }
-  }
-  const recoveryRows: Array<{ key: number; label: string; window: CapacityWindowView }> = showsAggregate && aggregation ? [
-    ...(aggregation.fiveHour ? [{ key: 0, label: t("codexAuth.fiveHour"), window: aggregation.fiveHour }] : []),
-    ...(aggregation.weekly ? [{ key: 1, label: t("codexAuth.weekly"), window: aggregation.weekly }] : []),
-    ...(aggregation.monthly ? [{ key: 2, label: t("codexAuth.monthly"), window: aggregation.monthly }] : []),
-    ...(aggregation.customWindows ?? []).map((window, index) => ({ key: index + 3, label: window.label, window })),
-  ] : [];
-  const formatPercent = (value: number) => new Intl.NumberFormat(locale, { maximumFractionDigits: 1 }).format(value);
-  const formatRecoveryAt = (value: number) => new Intl.DateTimeFormat(locale, {
-    dateStyle: "medium",
-    timeStyle: "short",
-  }).format(new Date(value > 10_000_000_000 ? value : value * 1000));
-
-  return (
-    <>
-      {showsAggregate && <div className="pws-capacity-label">{t("pws.capacity.estimate")}</div>}
-      {(primaryQuota || pending) && (
-        <QuotaBars
-          quota={primaryQuota}
-          threshold={80}
-          t={t}
-          layout="stacked"
-          pending={pending}
-          incompleteWindowKeys={showsAggregate ? incompleteWindowKeys : undefined}
-          incompleteCustomWindowLabels={showsAggregate ? incompleteCustomWindowLabels : undefined}
-        />
-      )}
-      {aggregation && (
-        <div className="pws-capacity-details">
-          {recoveryRows.flatMap(({ key, label, window }) => (
-            window.nextRecoveryAt !== undefined && window.nextRecoveryPercent !== undefined
-              ? [<div className="pws-capacity-recovery" key={key}>
-                  <span>{t("pws.capacity.nextRecovery")} · {label} · {formatRecoveryAt(window.nextRecoveryAt)}</span>
-                  <strong>{t("pws.capacity.recoveryShare", { percent: formatPercent(window.nextRecoveryPercent) })}</strong>
-                </div>]
-              : []
-          ))}
-          {showsAggregate && aggregation.currentAccount?.quota && (
-            <div className="pws-capacity-current">
-              <span className="pws-capacity-label">
-                {t("pws.capacity.currentAccount")}
-                {aggregation.currentAccount.plan ? ` · ${aggregation.currentAccount.plan}` : ""}
-              </span>
-              <QuotaBars quota={aggregation.currentAccount.quota} threshold={80} t={t} layout="stacked" />
-            </div>
-          )}
-          {aggregation.incomplete && aggregation.excludedAccounts > 0 && (
-            <div className="pws-capacity-incomplete">
-              {t("pws.capacity.incomplete", {
-                excluded: aggregation.excludedAccounts,
-                unknown: aggregation.unknownPlanAccounts,
-              })}
-            </div>
-          )}
-          {aggregation.partialWindowAccounts > 0 && (
-            <div className="pws-capacity-incomplete">
-              {t("pws.capacity.partial", { count: aggregation.partialWindowAccounts })}
-            </div>
-          )}
-        </div>
-      )}
-    </>
   );
 }
